@@ -1,11 +1,11 @@
-cmake_minimum_required(VERSION 3.10.2)
+cmake_minimum_required(VERSION 3.16.3)
 project(warehouse_ros)
 
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
 
-if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
+if(CMAKE_CXX_COMPILER_ID MATCHES "(GNU|Clang)")
   add_compile_options(-Wall -Wextra -Wpedantic)
 endif()
 
@@ -17,21 +17,10 @@ find_package(pluginlib REQUIRED)
 find_package(tf2 REQUIRED)
 find_package(tf2_ros REQUIRED)
 find_package(tf2_geometry_msgs REQUIRED)
-find_package(Boost REQUIRED COMPONENTS system filesystem thread)
+find_package(Boost REQUIRED)
 find_package(OpenSSL REQUIRED)
 
-include_directories(
-  include
-  ${OPENSSL_INCLUDE_DIR}
-)
-
-set(warehouse_srcs
-  src/database_loader.cpp
-  src/transform_collection.cpp)
-
-add_library(warehouse_ros SHARED ${warehouse_srcs})
-set_target_properties(warehouse_ros PROPERTIES WINDOWS_EXPORT_ALL_SYMBOLS TRUE)
-ament_target_dependencies(warehouse_ros
+set(THIS_PACKAGE_INCLUDE_DEPENDS
   rclcpp
   std_msgs
   geometry_msgs
@@ -39,37 +28,27 @@ ament_target_dependencies(warehouse_ros
   tf2
   tf2_ros
   tf2_geometry_msgs
+)
+
+add_library(warehouse_ros SHARED
+  src/database_loader.cpp
+  src/transform_collection.cpp
+)
+target_link_libraries(warehouse_ros PUBLIC OpenSSL::Crypto)
+set_target_properties(warehouse_ros PROPERTIES WINDOWS_EXPORT_ALL_SYMBOLS TRUE)
+ament_target_dependencies(warehouse_ros PUBLIC ${THIS_PACKAGE_INCLUDE_DEPENDS}
   Boost
 )
-target_link_libraries(warehouse_ros ${OPENSSL_CRYPTO_LIBRARY})
 target_include_directories(warehouse_ros
   PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
   PUBLIC $<INSTALL_INTERFACE:include>
 )
 
 add_executable(warehouse_test_dbloader src/test_dbloader.cpp)
-ament_target_dependencies(warehouse_test_dbloader
-  rclcpp
-  std_msgs
-  geometry_msgs
-  pluginlib
-  tf2
-  tf2_ros
-  tf2_geometry_msgs
+ament_target_dependencies(warehouse_test_dbloader PUBLIC ${THIS_PACKAGE_INCLUDE_DEPENDS}
   Boost
 )
-target_link_libraries(warehouse_test_dbloader warehouse_ros ${OPENSSL_CRYPTO_LIBRARY})
-
-ament_export_targets(export_${PROJECT_NAME} HAS_LIBRARY_TARGET)
-ament_export_dependencies(rclcpp)
-ament_export_dependencies(std_msgs)
-ament_export_dependencies(geometry_msgs)
-ament_export_dependencies(pluginlib)
-ament_export_dependencies(tf2)
-ament_export_dependencies(tf2_ros)
-ament_export_dependencies(tf2_geometry_msgs)
-ament_export_dependencies(Boost)
-ament_export_dependencies(OpenSSL)
+target_link_libraries(warehouse_test_dbloader PUBLIC warehouse_ros)
 
 install(
   TARGETS warehouse_ros
@@ -87,8 +66,14 @@ install(
 install(
   DIRECTORY include/
   DESTINATION include
-  FILES_MATCHING PATTERN "*.h" PATTERN "*.hpp"
 )
+
+install(DIRECTORY include/ DESTINATION include)
+
+ament_export_targets(export_${PROJECT_NAME} HAS_LIBRARY_TARGET)
+ament_export_dependencies(${THIS_PACKAGE_INCLUDE_DEPENDS})
+ament_export_dependencies(Boost)
+ament_export_dependencies(OpenSSL)
 
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
