@@ -30,33 +30,42 @@
 
 #include <warehouse_ros/database_loader.h>
 
-static const rclcpp::Logger LOGGER = rclcpp::get_logger("warehouse_ros.database_loader");
-
 namespace warehouse_ros
 {
-using std::string;
-
-DatabaseLoader::DatabaseLoader(const rclcpp::Node::SharedPtr& node) : node_(node)
+namespace
 {
-  initialize();
-}
+const rclcpp::Logger LOGGER = rclcpp::get_logger("warehouse_ros.database_loader");
 
-DatabaseLoader::~DatabaseLoader()
-{
-}
+// Parameter names
+const std::string WAREHOUSE_PLUGIN = "warehouse_plugin";
+const std::string WAREHOUSE_HOST = "warehouse_host";
+const std::string WAREHOUSE_PORT = "warehouse_port";
 
-void DatabaseLoader::initialize()
+// Default values
+const std::string WAREHOUSE_PLUGIN_DEFAULT = "warehouse_ros_mongo::MongoDatabaseConnection";
+const std::string WAREHOUSE_HOST_DEFAULT = "localhost";
+constexpr auto WAREHOUSE_PORT_DEFAULT = 33829;
+}  // namespace
+
+DatabaseLoader::DatabaseLoader(const rclcpp::node_interfaces::NodeParametersInterface::SharedPtr& node_parameters)
+  : node_parameters_(node_parameters)
 {
   // Create the plugin loader.
   try
   {
-    db_plugin_loader_.reset(new pluginlib::ClassLoader<DatabaseConnection>("warehouse_ros", "warehouse_ros::"
-                                                                                            "DatabaseConnection"));
+    db_plugin_loader_ = std::make_unique<pluginlib::ClassLoader<DatabaseConnection>>("warehouse_ros", "warehouse_ros::"
+                                                                                                      "DatabaseConnecti"
+                                                                                                      "on");
   }
   catch (pluginlib::PluginlibException& ex)
   {
     RCLCPP_FATAL_STREAM(LOGGER, "Exception while creating database_connection plugin loader " << ex.what());
   }
+}
+
+DatabaseLoader::DatabaseLoader(const rclcpp::Node::SharedPtr& node)
+  : DatabaseLoader(node->get_node_parameters_interface())
+{
 }
 
 typename DatabaseConnection::Ptr DatabaseLoader::loadDatabase()
@@ -68,49 +77,42 @@ typename DatabaseConnection::Ptr DatabaseLoader::loadDatabase()
 
   // Search for the warehouse_plugin parameter in the local namespace of the node, and up the tree of namespaces.
   // If the desired param is not found, make a final attempt to look for the param in the default namespace
-  string paramName;
-  // TODO: Revise parameter lookup
-  // if (!nh_.searchParam("warehouse_plugin", paramName))
-  paramName = "warehouse_plugin";
-  string db_plugin;
-  if (!node_->get_parameter_or(paramName, db_plugin, std::string("warehouse_ros_mongo::MongoDatabaseConnection")))
-    RCLCPP_ERROR(LOGGER, "Could not find parameter for database plugin name %s", db_plugin.c_str());
-
-  DatabaseConnection::Ptr db;
+  rclcpp::Parameter db_plugin(WAREHOUSE_PLUGIN, WAREHOUSE_PLUGIN_DEFAULT);
+  if (!node_parameters_->get_parameter(WAREHOUSE_PLUGIN, db_plugin))
+  {
+    RCLCPP_ERROR(LOGGER, "Could not find parameter '%s' using default '%s'", WAREHOUSE_PLUGIN.c_str(),
+                 db_plugin.as_string().c_str());
+  }
   try
   {
-    db = db_plugin_loader_->createUniqueInstance(db_plugin);
+    DatabaseConnection::Ptr db = db_plugin_loader_->createUniqueInstance(db_plugin.as_string());
+
+    // Get and set host name and port
+    rclcpp::Parameter host(WAREHOUSE_HOST, WAREHOUSE_HOST_DEFAULT);
+    if (!node_parameters_->get_parameter(WAREHOUSE_HOST, host))
+    {
+      RCLCPP_ERROR(LOGGER, "Could not find parameter '%s' using default '%s'", WAREHOUSE_HOST.c_str(),
+                   host.as_string().c_str());
+    }
+    rclcpp::Parameter port(WAREHOUSE_PORT, WAREHOUSE_PORT_DEFAULT);
+    if (!node_parameters_->get_parameter(WAREHOUSE_PORT, port))
+    {
+      RCLCPP_ERROR(LOGGER, "Could not find parameter '%s' using default '%li'", WAREHOUSE_PORT.c_str(), port.as_int());
+    }
+
+    // If successful return database pointer
+    if (db->setParams(host.as_string(), port.as_int()))
+    {
+      return db;
+    }
+    return typename DatabaseConnection::Ptr(new DBConnectionStub());
   }
   catch (pluginlib::PluginlibException& ex)
   {
-    RCLCPP_ERROR_STREAM(LOGGER,
-                        "Exception while loading database plugin '" << db_plugin << "': " << ex.what() << std::endl);
+    RCLCPP_ERROR_STREAM(LOGGER, "Exception while loading database plugin '" << db_plugin.as_string()
+                                                                            << "': " << ex.what() << std::endl);
     return typename DatabaseConnection::Ptr(new DBConnectionStub());
   }
-
-  bool hostFound = false;
-  bool portFound = false;
-
-  // TODO: Revise parameter lookup
-  // if (!nh_.searchParam("warehouse_host", paramName))
-  paramName = "warehouse_host";
-  std::string host;
-  node_->get_parameter_or(paramName, host, std::string("localhost"));
-  hostFound = true;
-
-  // TODO: Revise parameter lookup
-  // if (!nh_.searchParam("warehouse_port", paramName))
-  paramName = "warehouse_port";
-  int port;
-  node_->get_parameter_or(paramName, port, 33829);
-  portFound = true;
-
-  if (hostFound && portFound)
-  {
-    db->setParams(host, port);
-  }
-
-  return db;
 }
 
 MessageCollectionHelper::Ptr DBConnectionStub::openCollectionHelper(const std::string& /*db_name*/,
